refactor: moved reservation structs to dto
